Add helper "getPlainTag" + move it and "normalize" to the helpers folder
