+++ src/helpers/index.js
@@ -1,0 +1,4 @@
+import normalize from "./normalize";
+import getPlainTag from "./getPlainTag";
+
+export { normalize, getPlainTag }
