+++ src/helpers/normalize.js
